update examples
Revert "remove outdated workaround for svelte"

This reverts commit 31819ff472537e39019bf1cc3085cf0c61b3e0ca.

comment

publish svelte sdk fix

rename wrappers

rename wrappers svelte

simplify example

remove more

improve READMEs

remove css

fix

remove
