+++ examples/astro-solidjs/package.json
@@ -17,7 +17,7 @@
     "solid-js": "1.4.3"
   },
   "dependencies": {
-    "@builder.io/sdk-solid": "^0.0.8-11",
+    "@builder.io/sdk-solid": "^0.0.8-12",
     "solid-styled-components": "^0.27.6"
   }
 }
