+++ examples/svelte/sveltekit/README.md
@@ -1,6 +1,17 @@
 # Builder.io example with Sveltekit
 
-Builder.io drag-and-drop page and section building for Sveltekit!
+This example shows our [Svelte SDK](/packages/sdks/output/svelte) with Sveltekit.
+
+## Builder.io Setup
+
+- log into builder.io
+- from your account page, copy your API key and paste it into BUILDER_API_KEY in `DynamicallyRenderBuilderPage.svelte`
+- open the Builder.io Visual Editor for the model named "page"
+- enter http://localhost:3000 in the URL bar to the top right of the preview in Builder
+- drag a component into the layers tab, and it will appear in the Editor!
+
+Checkout this Loom for a visual walkthrough: https://www.loom.com/share/afd7c9a1f8f148959ea0396be42560fd (it's originally intended for React-Native, but all of the steps are still the exact same)
+
 
 ## Status
 
@@ -17,27 +28,3 @@ $ yarn run dev
 # or start the server and open the app in a new browser tab
 $ yarn run dev -- --open
 ```
-
-## Builder.io Setup
-
-- log into builder.io
-- from your account page, copy your API key and paste it into BUILDER_API_KEY in `[slug].svelte` and `index.svelte`
-- open the Builder.io Visual Editor for the model named "page"
-- enter http://localhost:3000 in the URL bar to the top right of the preview in Builder
-- drag a component into the layers tab, and it will appear in the Editor!
-
-Checkout this Loom for a visual walkthrough: https://www.loom.com/share/afd7c9a1f8f148959ea0396be42560fd (it's originally intended for React-Native, but all of the steps are still the exact same)
-
-For detailed explanation on how things work, check out [Nuxt.js docs](https://nuxtjs.org).
-
-## Building
-
-To create a production version of your app:
-
-```bash
-npm run build
-```
-
-You can preview the production build with `npm run preview`.
-
-> To deploy your app, you may need to install an [adapter](https://kit.svelte.dev/docs/adapters) for your target environment.
