--- examples/svelte/sveltekit/playwright.config.ts
@@ -1,10 +1,0 @@
-import type { PlaywrightTestConfig } from '@playwright/test';
-
-const config: PlaywrightTestConfig = {
-	webServer: {
-		command: 'npm run build && npm run preview',
-		port: 3000
-	}
-};
-
-export default config;
