--- examples/svelte/sveltekit/src/app.d.ts
@@ -1,15 +1,0 @@
-/// <reference types="@sveltejs/kit" />
-
-// See https://kit.svelte.dev/docs/types#the-app-namespace
-// for information about these interfaces
-declare namespace App {
-	interface Locals {
-		userid: string;
-	}
-
-	// interface Platform {}
-
-	// interface Session {}
-
-	// interface Stuff {}
-}
