+++ examples/svelte/sveltekit/src/lib/DynamicallyRenderBuilderPage.svelte
