+++ examples/svelte/sveltekit/src/routes/__layout.svelte
@@ -1,9 +1,6 @@
 <script lang="ts">
-	import Header from '$lib/header/Header.svelte';
 	import '../app.css';
 </script>
-
-<Header />
 
 <main>
 	<slot />
