+++ examples/vue/nuxt-2/README.md
@@ -2,7 +2,7 @@
 
 Example using Builder.io drag and drop page and section building for Nuxt using our beta v2 Vue SDK
 
-See [\_page.vue](pages/_page.vue) for rendering usage, and [register-builder-components.js](scripts/register-builder-components.js) for registering custom components
+See [BuilderContent.vue](./components/Builder/BuilderContent.vue) for rendering usage, and [init-builder.ts](./components/Builder/init-builder.ts) for registering custom components
 
 ## Status
