+++ examples/vue/nuxt-2/nuxt.config.js
@@ -1,15 +1,4 @@
 export default {
-  // Global page headers: https://go.nuxtjs.dev/config-head
-  head: {
-    title: 'nuxt',
-    meta: [
-      { charset: 'utf-8' },
-      { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      { hid: 'description', name: 'description', content: '' },
-    ],
-    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }],
-  },
-
   components: true,
 
   buildModules: ['@nuxt/components', '@builder.io/sdk-vue/nuxt'],
