+++ examples/vue/nuxt-2/pages/_page.vue
@@ -1,5 +1,5 @@
 <template>
-  <BuilderContent />
+  <BuilderDynamicallyRenderBuilderPage />
 </template>
 
 <script>
