+++ examples/vue/vue-storefront-2/README.md
@@ -1,6 +1,16 @@
 # Builder.io example with Vue Storefront
 
-Builder.io drag-and-drop page and section building for Vue Storefront
+This example shows our [Vue SDK](/packages/sdks/output/vue) with Vue Storefront.
+
+## Builder.io Setup
+
+- log into builder.io
+- from your account page, copy your API key and paste it into `BUILDER_API_KEY` in `DynamicallyRenderBuilderPage.vue`
+- open the Builder.io Visual Editor for the model named "page"
+- enter http://localhost:3000 in the URL bar to the top right of the preview in Builder
+- drag a component into the layers tab, and it will appear in the Editor!
+
+Checkout this Loom for a visual walkthrough: https://www.loom.com/share/afd7c9a1f8f148959ea0396be42560fd (it's originally intended for React-Native, but all of the steps are still the exact same)
 
 ## Status
 
@@ -24,15 +34,5 @@ $ yarn start
 # generate static project
 $ yarn generate
 ```
-
-## Builder.io Setup
-
-- log into builder.io
-- from your account page, copy your API key and paste it into BUILDER_API_KEY in Home.vue
-- open the Builder.io Visual Editor for the model named "page"
-- enter http://localhost:3000 in the URL bar to the top right of the preview in Builder
-- drag a component into the layers tab, and it will appear in the Editor!
-
-Checkout this Loom for a visual walkthrough: https://www.loom.com/share/afd7c9a1f8f148959ea0396be42560fd (it's originally intended for React-Native, but all of the steps are still the exact same)
 
 For detailed explanation on how things work, check out [Nuxt.js docs](https://nuxtjs.org).
