+++ examples/vue/vue-storefront-2/pages/_path.vue
@@ -1,6 +1,6 @@
 <template>
   <div id="home">
-    <BuilderContent />
+    <DynamicallyRenderBuilderPage />
   </div>
 </template>
 <script>
